chore: Update next.config.mjs and page.tsx
Refactor next.config.mjs to conditionally set the basePath based on the NODE_ENV environment variable. This improves the configuration for the production environment by setting the basePath to '/mirko-portfolio'.

In page.tsx, update the getAllPosts function to fetch posts from both the 'content' and 'blog' directories. This ensures that all relevant posts are retrieved.